+++ tests/tests_unit/test_data_classes/test_capabilities.py
@@ -12,6 +12,7 @@ from cognite.client.data_classes.capabilities import (
     AllScope,
     Capability,
     CurrentUserScope,
+    DataModelInstancesAcl,
     DataSetsAcl,
     EventsAcl,
     ExperimentsAcl,
@@ -20,6 +21,7 @@ from cognite.client.data_classes.capabilities import (
     ProjectCapabilityList,
     ProjectsAcl,
     RawAcl,
+    SpaceIDScope,
     TableScope,
     UnknownAcl,
     UnknownScope,
@@ -246,6 +248,39 @@ class TestCapabilities:
             Capability.load(dumped)
 
         assert Capability.load(dumped, allow_unknown=True)
+
+    @pytest.mark.parametrize("has_write_allscope", (True, False))
+    @pytest.mark.parametrize("has_write_props_allscope", (True, False))
+    @pytest.mark.parametrize("has_write_in_same_scope", (True, False))
+    def test_load_data_model_instances__with_write_properties(
+        self, cognite_client, has_write_allscope, has_write_props_allscope, has_write_in_same_scope
+    ):
+        # WRITE_PROPERTIES grants a subset of capabilities of WRITE, so we ensure that having just WRITE
+        # won't cause WRITE_PROPERTIES to be reported as missing.
+        existing = [
+            DataModelInstancesAcl([DataModelInstancesAcl.Action.Write], SpaceIDScope(["foo", "this"])),
+            DataModelInstancesAcl([DataModelInstancesAcl.Action.Write_Properties], SpaceIDScope(["bar"])),
+        ]
+        if has_write_allscope:
+            existing.append(DataModelInstancesAcl([DataModelInstancesAcl.Action.Write], AllScope()))
+        if has_write_in_same_scope:
+            existing.append(DataModelInstancesAcl([DataModelInstancesAcl.Action.Write], SpaceIDScope(["too_much"])))
+        if has_write_props_allscope:
+            existing.append(DataModelInstancesAcl([DataModelInstancesAcl.Action.Write_Properties], AllScope()))
+
+        desired = [
+            DataModelInstancesAcl([DataModelInstancesAcl.Action.Write_Properties], SpaceIDScope(["foo", "too_much"])),
+            DataModelInstancesAcl([DataModelInstancesAcl.Action.Write_Properties], SpaceIDScope(["bar"])),
+        ]
+        if has_write_allscope or has_write_props_allscope or has_write_in_same_scope:
+            expected_missing = []
+        else:
+            expected_missing = [
+                DataModelInstancesAcl([DataModelInstancesAcl.Action.Write_Properties], SpaceIDScope(["too_much"]))
+            ]
+
+        missing = cognite_client.iam.compare_capabilities(existing_capabilities=existing, desired_capabilities=desired)
+        assert missing == expected_missing
 
     def test_create_capability_forget_initializing_scope(self):
         # Ensure these do not raise. All other scopes require arguments and so will